backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
passlib[bcrypt]==1.7.4
pytest==8.3.2
ruff==0.6.2
black==24.8.0
isort==5.13.2
